+++ esbuild.mjs
@@ -55,7 +55,8 @@ const builds = [
       "test/generator.test.ts",
       "test/streaming.test.ts",
       "test/outline-generator.test.ts",
-      "test/targets.test.ts"
+      "test/targets.test.ts",
+      "test/manifest.test.ts"
     ],
     outdir: "dist/test",
     entryNames: "[name]",